refactor : 난잡해진 폴더 구조 정리
- 도메인 별로 나누어지지 않은 패키지(validator, util, model, aop 등)를 공통 폴더로 이동
- 사용하지 않는 DBServerException 삭제(모두 spring 표준 Exception 대체하였음)
