+++ src/main/java/gift/common/model/CustomPage.java
@@ -1,4 +1,4 @@
-package gift.model;
+package gift.common.model;
 
 import java.util.List;
 import java.util.function.Function;
