+++ src/main/java/gift/common/validation/group/AuthenticationGroups.java
@@ -1,4 +1,4 @@
-package gift.validation.group;
+package gift.common.validation.group;
 
 public class AuthenticationGroups {
     public interface UserGroup {}
