+++ src/main/java/gift/common/validation/validator/KakaoNotContainedValidator.java
@@ -1,6 +1,6 @@
-package gift.validation.validator;
+package gift.common.validation.validator;
 
-import gift.validation.annotation.KakaoNotContained;
+import gift.common.validation.annotation.KakaoNotContained;
 import jakarta.validation.ConstraintValidator;
 import jakarta.validation.ConstraintValidatorContext;
