--- src/main/java/gift/exception/DBServerException.java
@@ -1,10 +1,0 @@
-package gift.exception;
-
-public class DBServerException extends RuntimeException {
-    public DBServerException(String message) {
-        super(message);
-    }
-    public DBServerException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
