+++ src/main/java/gift/repository/product/ProductRepository.java
@@ -1,6 +1,6 @@
 package gift.repository.product;
 
-import gift.model.CustomPage;
+import gift.common.model.CustomPage;
 import gift.entity.Product;
 
 import java.util.List;
