+++ src/main/resources/application.properties
@@ -18,3 +18,7 @@ server.error.path=/error
 
 # logging
 logging.level.org.springframework.web=DEBUG
+
+# jwt
+gift.jwt.secret=b12277e4d918a0d95f2f5adc476cd3bff18911eb3d8f34fb5067f78326c5c6f0
+gift.jwt.expiration=3600
